requirements.txt
# Web 框架
flask>=3.0.0
flask-cors>=4.0.0

# 数据库
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0

# 认证
Flask-JWT-Extended>=4.6.0

# 工具
werkzeug>=3.0.0
python-dotenv>=1.0.0

# 向量数据库
chromadb>=0.4.0